Explicitly exit the process to not wait for hanging promises
As with other actions like setup-node, I'm seeing 2-4 minute delays in post cache actions lately. Apparently this is because of a change in node behavior: https://github.com/ruby/setup-ruby/issues/543#issuecomment-1793608370

The fix, as with other actions, is to explicitly exit so as not to wait for hanging promises.